refactor(web): migrate events into src/domains
Second pilot for the domain-driven structure. Co-locate the calendar-events
feature under apps/web/src/domains/events:
- components/ <- components/events/* (RSVP, attendance, delete buttons, banners)
- server/ <- lib/events/* (create/update/delete, permissions, recurrence,
  recurring-operations, labels, event-type-options)
- index.ts barrel + README; repoint all importers at @/domains/events

Deliberately left in shared infra (documented in the domain README): event
validation schemas (lib/schemas/content + events-ai, shared with the AI
subsystem and announcements), calendar sync/provider mappers, analytics and
Stripe "event" modules, and the AI tool registry definitions. Typecheck passes.
