+++ apps/web/src/app/[orgSlug]/calendar/events/[eventId]/edit/page.tsx
@@ -9,9 +9,7 @@ import { Card, Button, Input, Select, Textarea } from "@/components/ui";
 import { PageHeader } from "@/components/layout";
 import { useLocale, useTranslations } from "next-intl";
 import { editEventSchema, type EditEventForm } from "@/lib/schemas/content";
-import { updateFutureEvents } from "@/lib/events/recurring-operations";
-import { resolveEventActionLabel } from "@/lib/events/labels";
-import { EVENT_TYPE_OPTIONS } from "@/lib/events/event-type-options";
+import { updateFutureEvents, resolveEventActionLabel, EVENT_TYPE_OPTIONS } from "@/domains/events";
 import { calendarEventDetailPath } from "@/lib/calendar/routes";
 import { localToUtcIso, utcToLocalParts, resolveOrgTimezone } from "@/lib/utils/timezone";
 import type { NavConfig } from "@/lib/navigation/nav-items";
