+++ apps/web/src/app/api/ai/[orgId]/pending-actions/[actionId]/confirm/handler.ts
@@ -32,9 +32,7 @@ import {
   supportsDraftSessionsStore,
   type DraftSessionSupabase,
 } from "@/lib/ai/draft-sessions";
-import { createEvent } from "@/lib/events/create-event";
-import { updateEvent } from "@/lib/events/update-event";
-import { deleteEvent } from "@/lib/events/delete-event";
+import { createEvent, updateEvent, deleteEvent } from "@/domains/events";
 import { createJobPosting } from "@/lib/jobs/create-job";
 import { updateJobPosting } from "@/lib/jobs/update-job";
 import { deleteJobPosting } from "@/lib/jobs/delete-job";
