+++ apps/web/src/domains/events/components/AttendanceList.tsx
