+++ apps/web/src/domains/events/components/EventDeleteButton.tsx
