+++ apps/web/src/domains/events/components/EventRsvp.tsx
