+++ apps/web/src/domains/events/components/index.ts
