+++ apps/web/src/domains/events/server/delete-event.ts
