+++ apps/web/src/domains/events/server/permissions.ts
